+++ CHANGELOG.md
@@ -23,6 +23,7 @@ Please avoid adding duplicate information across this changelog and JIRA/doc inp
 - UI: Hovering over a node in the Network Graph will show that node's listening ports (ROX-5469)
 - Fixed an issue on the API docs page where the left menu panel no longer scrolled independently of the main content.
 - UI: Added `Scanner` to the image single page in Vuln Mgmt (ROX-5289)
+- In `v1/clusters` response, `status.lastContact` has been deprecated, hence use `healthStatus.lastContact` instead.
 - UI: Disable the Next button when required fields are empty in the Cluster form (ROX-5519)
 
 ## [48.0]